refactor(settings): rename to settings manager
