+++ src/bot/managers/SettingsManager.ts
